requirements.txt
asttokens==2.4.1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
executing==2.0.1
filelock==3.13.1
Flask==3.0.3
fonttools==4.54.1
fsspec==2024.2.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jupyter_core==5.7.1
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.6
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.3
parso==0.8.3
pillow==10.2.0
prompt-toolkit==3.0.43
psutil==6.1.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
Pygments==2.17.2
pyparsing==3.2.0
python-dateutil==2.8.2
pytz==2024.2
PyYAML==6.0.2
pyzmq==25.1.2
requests==2.32.3
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
stack-data==0.6.3
sympy==1.13.1
torch
torchaudio
torchvision
tornado==6.4
tqdm==4.67.0
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2024.2
ultralytics==8.3.28
ultralytics-thop==2.0.10
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.1.2
mediapipe
